still problems with the I2Cdev library

+++ Project/Inc/I2Cdev.h
@@ -55,7 +55,7 @@ public:
         int8_t readByte(uint8_t devAddr, uint8_t regAddr, uint8_t *data);
         int8_t readWord(uint8_t devAddr, uint8_t regAddr, uint16_t *data);
         int8_t readBytes(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint8_t *data);
-        int8_t readBytes_16bitaddr(uint8_t devAddr, uint16_t regAddr, uint8_t length, uint8_t *data);
+        int8_t readBytes_16bitaddr(uint8_t devAddr, uint16_t regAddr, uint32_t length, uint8_t *data);
         int8_t readWords(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint16_t *data);
 
         bool writeBit(uint8_t devAddr, uint8_t regAddr, uint8_t bitNum, uint8_t data);
@@ -65,7 +65,7 @@ public:
         bool writeByte(uint8_t devAddr, uint8_t regAddr, uint8_t data);
         bool writeWord(uint8_t devAddr, uint8_t regAddr, uint16_t data);
         bool writeBytes(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint8_t *data);
-        bool writeBytes_16bitaddr(uint8_t devAddr, uint16_t regAddr, uint8_t length, uint8_t *data);
+        bool writeBytes_16bitaddr(uint8_t devAddr, uint16_t regAddr, uint32_t length, uint8_t *data);
         bool writeWords(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint16_t *data);
 };
